Revert 'Merge pull request #6008 from GoogleCloudPlatform/revert-5773-e2e_kubectl_util' with additional clean up, disambiguate the container name from the json contents.
